chore: 收口 v9.1/3O 内化/H3/导演台等工作区并修复 CI 门禁
工作区收口(90+ 文件,含此前未提交的整批新模块):
- v9.1 任务大盘:TaskRun SQLite 状态机 + WorkspaceManager 断点续跑 +
  dashboard API(列表/详情/成片输出/状态计数)+ WebSocket 推流(ping/pong)
- 3O 内化:embrace runtime/failure registry/replay trace/aesthetic canon/
  recipe schema + canon numbering CI 检查
- H3 本地供给:ComfyUI 客户端 + h3_compiler + post(FlashVSR/RIFE)+ 路由
- history_series:教材↔古籍交叉弧 + 每日连载产线(修复 animate_lesson 漏传
  textbook_text、%%s 日志、重复 _fail 三个真 bug)
- pipeline_lite(Playwright 录屏/ASR)、freevideo、scene_block、pr_to_video、
  music_to_video、promo_video 等 assembly workflow
- cinematic 黄金公式动画、director 草图/图像池/修复 agent、audio 多模型路由
- gen-engine 服务(vibeasr-src vendor,ruff exclude 已存在)、hevi skills CLI

修复使 CI 门禁全绿(main 此前已红):
- ruff:main 307 错 → 0(含 per-file E501 豁免,同 freevideo 先例)
- mypy --strict:189 错 → 0(443 files;新增 asyncpg/dashscope/cv2/oservi
  ignore_missing_imports + oprim no-untyped-call 放行)
- 5 个 3O CI 检查全过(删除未用的 run_free_video_workflow)
- pytest 1853 全过(修复:node_mapper 死代码转 plan-only、dashboard 空 stub
  补全实现、CLIP 纯色兜底改固定 seed 噪声图、API-CAPABILITIES.md 重新生成)
- hevi-web tsc + vitest 84/84;hevi-remotion tsc 全过

其他:output/tasks 由容器 root 写入导致测试 PermissionError,已移出
(output/tasks.container-root 保留数据)并重建 soffy 所有目录。

+++ hevi/tongjian/voiceover.py
@@ -107,7 +107,7 @@ async def _get_audio_duration_ms(path: Path) -> int:
 
     try:
         duration_s = await asyncio.to_thread(probe_duration, path)
-        return round(duration_s * 1000)
+        return int(round(duration_s * 1000))
     except Exception as e:
         logger.warning("ffprobe 读取时长失败 (%s),按文件大小估算: %s", path, e)
         # 粗略估算:16kHz 16bit mono WAV ≈ 32KB/s
